# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: 优刻（YOUC）平台 API
  version: 0.19.0-unreleased
  description: |
    优刻全站接口：用户面统一前缀 `/api/v1`，管理面为 `/api/admin`（不进公开版本契约）。

    鉴权：
    - 业务接口走 `Authorization: Bearer <accessToken>`（HS256 JWT，15 分钟有效，登录/刷新接口下发）。
    - 刷新令牌走 HttpOnly cookie `ic_refresh`（Path=/api，30 天，单次轮换，服务端只存哈希）。
    - 媒体读取额外接受 HttpOnly cookie `ic_media`（Path=/api/v1/media，为 `<img>/<video>` 直链准备的只读凭据；
      `/api/v1/media` 与 `/api/v1/media-download` 的 GET/HEAD 可用 Bearer 或该 cookie，写操作仅认 Bearer）。

    全局约定：
    - 统一错误信封 `{"error":{"code":"...","message":"...","fields":{...}}}`（fields 仅字段级错误时出现，
      error 对象可能平铺附加键，如 REVISION_CONFLICT 附 revision）；限流响应附 `Retry-After` 头。
    - 时间一律 RFC3339Nano UTC 字符串；金额单位为微元（1 元 = 1000000 micros）。
    - 分页：page/size（page 从 1 起、size 默认 20 上限 100）或游标（cursor + nextCursor）两种，按接口标注。
    - 站点维护模式开启时，非管理员写请求一律 503 MAINTENANCE_MODE（认证、OIDC、管理面、支付回调豁免）。
    - 强制改密账号仅可调用登出、刷新与修改密码，其余接口 403 PASSWORD_CHANGE_REQUIRED。
    - 通用错误（除特别标注外各组适用）：401 UNAUTHORIZED / TOKEN_EXPIRED、403 ACCOUNT_DISABLED、
      403 PASSWORD_CHANGE_REQUIRED、503 MAINTENANCE_MODE、500 INTERNAL_ERROR。
servers:
  - url: https://sim-art.youc.online
    description: 测试环境
  - url: https://youc.online
    description: 正式环境（主站应用部署后同域；当前该域为静态官网）
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录/刷新接口返回的 accessToken（HS256，15 分钟）。
    mediaCookie:
      type: apiKey
      in: cookie
      name: ic_media
      description: 媒体只读 cookie（Path=/api/v1/media，HttpOnly），登录/刷新时下发轮换；仅 GET/HEAD 可代替 Bearer。
...
paths:
...
  /api/v1/me:
    get:
      tags: [用户]
      summary: 当前用户聚合档案
      responses:
        '200':
          description: 用户、档位、点数、用量、媒体到期与注销状态聚合
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MeResponse' }
        '401': { $ref: '#/components/responses/Error401' }
    patch:
      tags: [用户]
      summary: 更新资料（昵称/头像）
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                displayName: { type: string, nullable: true }
                avatarUrl: { type: string, nullable: true }
      responses:
        '200':
          description: 更新后的用户
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/UserPayload' }
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
  /api/v1/me/password:
    post:
      tags: [用户]
      summary: 修改密码（强制改密白名单出口）
      description: '成功后撤销全部旧会话、轮换两枚 cookie、媒体令牌版本 +1；响应头 X-Session-Refreshed: true。主要错误：401 INVALID_CREDENTIALS（旧密码错）。'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oldPassword, newPassword]
              properties:
                oldPassword: { type: string }
                newPassword: { type: string, description: 8-72 位 }
      responses:
        '204': { description: 已修改 }
        '401': { $ref: '#/components/responses/Error401' }
  /api/v1/me/free-grant/claim:
    post:
      tags: [用户]
      summary: 领取免费体验额度（幂等）
      description: 主要错误：403 FREE_GRANT_UNAVAILABLE（开关关闭/风控/日预算耗尽）、403 EMAIL_NOT_VERIFIED、403 ACCOUNT_PENDING_DELETION。
      responses:
        '200':
          description: 领取结果（已领取返回既有记录）
          content:
            application/json:
              schema:
                type: object
                properties:
                  campaignId: { type: string }
                  status: { type: string }
                  imageTrials: { type: integer }
                  videoTrials: { type: integer }
                  grantedMicros: { type: integer, format: int64 }
        '403': { $ref: '#/components/responses/Error403' }
  /api/v1/me/deletion:
    post:
      tags: [用户]
      summary: 申请注销（7 天冷静期，重复申请幂等）
      description: 申请后立即无法生成与充值，仍可浏览与导出。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password: { type: string, description: 密码确认 }
      responses:
        '200':
          description: 已进入冷静期
          content:
            application/json:
              schema:
                type: object
                properties:
                  scheduledAt: { type: string, description: 到期匿名化时间 }
        '401': { $ref: '#/components/responses/Error401' }
  /api/v1/me/deletion/cancel:
    post:
      tags: [用户]
      summary: 撤销注销
      responses:
        '200':
          description: 账号恢复
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string }
        '409': { $ref: '#/components/responses/Error409', description: 不在冷静期 DELETION_NOT_PENDING }
  /api/v1/me/export:
    get:
      tags: [用户]
      summary: 个人数据导出（JSON 附件）
      description: 'Content-Disposition: attachment; filename="youc-export.json"；含资料/点数/订单/生成记录/画布元数据，不含媒体二进制。'
      responses:
        '200':
          description: 导出 JSON
        '401': { $ref: '#/components/responses/Error401' }
  /api/v1/oidc/authorize:
    get:
      tags: [OIDC]
      summary: 授权端点（授权码 + 强制 PKCE S256）
      description: |
        浏览器直跳 302 到 redirect_uri?code=...&state=...；请求头 Accept 含 application/json 时返回 JSON 模式
        {"redirect": "..."} 由前端自行跳转。code 一次性、10 分钟有效。client/redirect_uri/PKCE 校验失败一律
        400 JSON 且绝不重定向。IP 限流 20 次/10 分钟。
      parameters:
        - { name: client_id, in: query, required: true, schema: { type: string } }
        - { name: redirect_uri, in: query, required: true, schema: { type: string }, description: 须与注册完全一致 }
        - { name: response_type, in: query, required: true, schema: { type: string, enum: [code] } }
        - { name: scope, in: query, required: true, schema: { type: string }, description: 须含 openid }
        - { name: code_challenge, in: query, required: true, schema: { type: string } }
        - { name: code_challenge_method, in: query, required: true, schema: { type: string, enum: [S256] } }
        - { name: state, in: query, required: false, schema: { type: string }, description: 原样回传 }
      responses:
        '302':
          description: 浏览器直跳模式，重定向到 redirect_uri
        '200':
          description: 'JSON 模式（Accept: application/json）'
          content:
            application/json:
              schema:
                type: object
                properties:
                  redirect: { type: string, description: 完整回跳 URL（带 code 与 state） }
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
        '429': { $ref: '#/components/responses/Error429' }
  /api/v1/oidc/token:
    post:
      tags: [OIDC]
      summary: 授权码换 token（RS256，15 分钟）
      description: '客户端凭 client_secret 认证（恒时比对）；JSON 或 form 皆可。主要错误：401 INVALID_CLIENT、400 INVALID_GRANT（code 无效/已用/过期、redirect_uri 不匹配、verifier 错）、400 UNSUPPORTED_GRANT_TYPE。IP 限流 20 次/10 分钟。'
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [grant_type, client_id, client_secret, code, redirect_uri, code_verifier]
              properties:
                grant_type: { type: string, enum: [authorization_code] }
                client_id: { type: string }
                client_secret: { type: string }
                code: { type: string }
                redirect_uri: { type: string }
                code_verifier: { type: string }
      responses:
        '200':
          description: 签发成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token: { type: string }
                  token_type: { type: string }
                  expires_in: { type: integer, description: 900 秒 }
                  scope: { type: string }
                  id_token: { type: string, description: RS256 JWT，claims 仅 sub/aud }
        '400': { $ref: '#/components/responses/Error400' }
        '401': { $ref: '#/components/responses/Error401' }
  /api/v1/oidc/userinfo:
    get:
      tags: [OIDC]
      summary: OIDC 用户信息
      description: 验 OIDC 签发的 RS256 access token，与平台会话令牌是两套凭据，平台 Bearer 会被拒绝。
      parameters:
        - { name: Authorization, in: header, required: true, schema: { type: string }, description: 'Bearer <OIDC access_token>' }
      responses:
        '200':
          description: 用户档案
          content:
            application/json:
              schema:
                type: object
                properties:
                  sub: { type: string }
                  email: { type: string }
                  username: { type: string }
                  displayName: { type: string, nullable: true }
                  emailVerified: { type: boolean }
        '401': { $ref: '#/components/responses/Error401' }
...
  /api/v1/canvases:
    get:
      tags: [画布]
      summary: 我的画布列表（服务端分页/搜索/排序，只回摘要）
      parameters:
        - { $ref: '#/components/parameters/PageParam' }
        - { $ref: '#/components/parameters/SizeParam' }
        - { name: sort, in: query, schema: { type: string }, description: '白名单 updatedAt|createdAt|title，前缀 - 倒序，默认 -updatedAt' }
        - { name: q, in: query, schema: { type: string }, description: 标题关键字 }
      responses:
        '200':
          description: 画布摘要分页
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/CanvasSummary' } }
                  total: { type: integer, format: int64 }
                  page: { type: integer }
                  size: { type: integer }
        '400': { $ref: '#/components/responses/Error400' }
    post:
      tags: [画布]
      summary: 创建画布
      description: data 顶层键 nodes[]/connections[]，服务端据此计算计数与封面，上限 2MB。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, maxLength: 200 }
                data: { type: object }
      responses:
        '201':
          description: 画布详情（revision=1）
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CanvasDetail' }
        '400': { $ref: '#/components/responses/Error400' }
...
  /api/v1/assets/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [素材]
      summary: 素材详情
      responses:
        '200':
          description: 素材对象
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetItem' }
        '404': { $ref: '#/components/responses/Error404' }
    patch:
      tags: [素材]
      summary: 部分更新素材
      description: 至少提供一个字段；storageKey/bytes 更新时自动镜像写进 data。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, maxLength: 200 }
                tags: { type: array, items: { type: string } }
                data: { type: object }
                storageKey: { type: string }
                bytes: { type: integer, format: int64 }
      responses:
        '200':
          description: 更新后的素材
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetItem' }
        '404': { $ref: '#/components/responses/Error404' }
    delete:
      tags: [素材]
      summary: 删除素材（幂等）
      responses:
        '204': { description: 已删除 }
        '404': { $ref: '#/components/responses/Error404' }
...
  /api/v1/generations/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [生成记录]
      summary: 生成记录详情
      responses:
        '200':
          description: 生成记录对象
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GenerationItem' }
        '404': { $ref: '#/components/responses/Error404' }
    delete:
      tags: [生成记录]
      summary: 删除生成记录（幂等）
      responses:
        '204': { description: 已删除 }
        '404': { $ref: '#/components/responses/Error404' }
...
  /api/admin/me:
    get:
      tags: [Admin-元信息]
      summary: 后台当前用户（免权限点，需后台角色）
      responses:
        '200': { description: 当前管理员信息 }
        '403': { $ref: '#/components/responses/Error403' }
...